requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyPDF2
pdfplumber
pytesseract
Pillow
pandas
python-docx
openpyxl
numpy
scikit-learn
nltk
spacy
flask
watchdog
imutils
opencv-python
